Move path separator constant to common
