+++ AdminServer/appscale/admin/constants.py
@@ -81,9 +81,6 @@ IMMUTABLE_PROJECTS = [DASHBOARD_APP_ID]
 # The directory where source archives are stored.
 SOURCES_DIRECTORY = os.path.join('/', 'opt', 'appscale', 'apps')
 
-# The character used to separate portions of a complete version string.
-VERSION_PATH_SEPARATOR = '_'
-
 # The ZooKeeper location for storing version details.
 VERSION_NODE_TEMPLATE = ('/appscale/projects/{project_id}'
                          '/services/{service_id}/versions/{version_id}')
