+++ AdminServer/appscale/admin/instance_manager/projects_manager.py
@@ -7,7 +7,7 @@ import os
 from tornado.ioloop import IOLoop
 
 from appscale.common.constants import CONFIG_DIR
-from ..constants import VERSION_PATH_SEPARATOR
+from appscale.common.constants import VERSION_PATH_SEPARATOR
 
 
 class VersionManager(object):
