+++ common/appscale/common/constants.py
@@ -141,3 +141,6 @@ SMALL_WAIT = 5
 
 # The number of seconds to wait before retrying some operations.
 TINY_WAIT = .1
+
+# The character used to separate portions of a complete version string.
+VERSION_PATH_SEPARATOR = '_'
